% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase.R
\name{coinbase_account_info}
\alias{coinbase_account_info}
\title{Coinbase Account Info Function}
\usage{
coinbase_account_info(key, secret_key)
}
\arguments{
\item{key}{Your coinbase API key}

\item{secret_key}{Your coinbase secret API key}
}
\value{
A list of all Coinbase accounts with details of their holdings.
}
\description{
A function for accessing account information via the Coinbase API.
}
