% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase_accounts.R
\name{coinbase_accounts}
\alias{coinbase_accounts}
\title{Coinbase Account Info Function}
\usage{
coinbase_accounts(key, secret_key)
}
\arguments{
\item{key}{Your coinbase API key}

\item{secret_key}{Your coinbase secret API key}
}
\value{
A list of all Coinbase accounts with details of their holdings.
}
\description{
A function for accessing account information via the Coinbase API.
}
