% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase_pro_account_history.R
\name{coinbase_pro_account_history}
\alias{coinbase_pro_account_history}
\title{Coinbase Pro Account Detail}
\usage{
coinbase_pro_account_history(key, secret_key, passphrase, account)
}
\arguments{
\item{key}{Your Coinbase Pro API key}

\item{secret_key}{Your Coinbase Pro secret API key}

\item{passphrase}{Your Coinbase Pro API passphrase}

\item{account}{The Coinbase Pro account ID you wish to retreive history for}
}
\value{
A list of all Coinbase Pro accounts with details of their holdings.
}
\description{
A function for accessing account information via the Coinbase Pro API.
}
