% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase_pro_accounts.R
\name{coinbase_pro_accounts}
\alias{coinbase_pro_accounts}
\title{Coinbase Pro Account Listing Function}
\usage{
coinbase_pro_accounts(key, secret_key, passphrase)
}
\arguments{
\item{key}{Your Coinbase Pro API key}

\item{secret_key}{Your Coinbase Pro secret API key}

\item{passphrase}{Your Coinbase Pro API passphrase}
}
\value{
A list of all Coinbase Pro accounts with details of their holdings.
}
\description{
A function for accessing account information via the Coinbase Pro API.
}
